Use app token for conversation labeling
